update controller package
