+++ app/src/main/java/hexlet/code/controller/AuthenticationController.java
@@ -1,0 +1,31 @@
+package hexlet.code.controller;
+
+import hexlet.code.dto.AuthRequest;
+import hexlet.code.util.JWTUtils;
+import lombok.AllArgsConstructor;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api")
+@AllArgsConstructor
+public class AuthenticationController {
+
+    private JWTUtils jwtUtils;
+    private AuthenticationManager authenticationManager;
+
+    @PostMapping("/login")
+    public String create(@RequestBody AuthRequest authRequest) {
+        var authentication = new UsernamePasswordAuthenticationToken(
+                authRequest.getUsername(),
+                authRequest.getPassword());
+
+        authenticationManager.authenticate(authentication);
+
+        return jwtUtils.generateToken(authRequest.getUsername());
+    }
+}
