--- app/src/main/java/hexlet/code/controller/api/AuthenticationController.java
@@ -1,38 +1,0 @@
-package hexlet.code.controller.api;
-
-import hexlet.code.dto.AuthRequest;
-import hexlet.code.util.JWTUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api")
-public class AuthenticationController {
-
-    @Autowired
-    private JWTUtils jwtUtils;
-
-    @Autowired
-    private AuthenticationManager authenticationManager;
-
-    @PostMapping("/login")
-    public ResponseEntity<String> create(@RequestBody AuthRequest authRequest) {
-        try {
-            var authentication = new UsernamePasswordAuthenticationToken(
-                    authRequest.getUsername(), authRequest.getPassword());
-
-            authenticationManager.authenticate(authentication);
-
-            return ResponseEntity.ok(jwtUtils.generateToken(authRequest.getUsername()));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-    }
-}
