--- app/src/main/java/hexlet/code/controller/api/UserController.java
@@ -1,74 +1,0 @@
-package hexlet.code.controller.api;
-
-import hexlet.code.dto.UserDTO.UserCreateDTO;
-import hexlet.code.dto.UserDTO.UserDTO;
-import hexlet.code.dto.UserDTO.UserUpdateDTO;
-import hexlet.code.repository.UserRepository;
-import hexlet.code.service.UserService;
-import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/users")
-public final class UserController {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private UserService userService;
-
-    @GetMapping(path = "/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<UserDTO> show(@PathVariable Long id) {
-        var user = userService.findById(id);
-        return ResponseEntity.ok()
-                .body(user);
-    }
-
-    @GetMapping(path = "")
-    @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<List<UserDTO>> index() {
-        var users = userService.getAllUsers();
-        return ResponseEntity.ok()
-                .header("X-Total-Count", String.valueOf(users.size()))
-                .body(users);
-    }
-
-    @PostMapping(path = "")
-    @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<UserDTO> create(@Valid @RequestBody UserCreateDTO data) throws URISyntaxException {
-        var user = userService.createUser(data);
-        return ResponseEntity.created(new URI("/api/users/" + user.getId()))
-                .body(user);
-    }
-
-    @PutMapping(path = "/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<UserDTO> update(@Valid @RequestBody UserUpdateDTO data, @PathVariable Long id) {
-        var user = userService.updateUser(id, data);
-        return ResponseEntity.ok()
-                .body(user);
-    }
-
-    @DeleteMapping(path = "/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    void delete(@PathVariable Long id) {
-        userService.deleteUser(id);
-    }
-}
